+++ lib/xandra/cluster/pool.ex
@@ -7,11 +7,12 @@ defmodule Xandra.Cluster.Pool do
   # to work on the internals. Plus, the surface API of *this* internal module is
   # quite small, so it's really mostly internals.
 
-  use GenServer
+  @behaviour :gen_statem
 
   alias Xandra.Cluster.{Host, LoadBalancingPolicy}
 
-  @genserver_opts [:debug, :hibernate_after, :spawn_opt]
+  @genstatem_opts [:debug, :hibernate_after, :spawn_opt]
+  @state :no_state
 
   ## Public API
 
@@ -20,12 +21,16 @@ defmodule Xandra.Cluster.Pool do
     {sync_connect_timeout, cluster_opts} = Keyword.pop!(cluster_opts, :sync_connect)
 
     # Split out GenServer-specific options from the cluster options.
-    {genserver_opts, cluster_opts} = Keyword.split(cluster_opts, @genserver_opts)
-    genserver_opts = Keyword.merge(genserver_opts, Keyword.take(cluster_opts, [:name]))
+    {genstatem_opts, cluster_opts} = Keyword.split(cluster_opts, @genstatem_opts)
+    genstatem_opts = Keyword.merge(genstatem_opts, Keyword.take(cluster_opts, [:name]))
 
     alias_or_nil = if sync_connect_timeout, do: alias()
 
-    case GenServer.start_link(__MODULE__, {cluster_opts, pool_opts, alias_or_nil}, genserver_opts) do
+    case :gen_statem.start_link(
+           __MODULE__,
+           {cluster_opts, pool_opts, alias_or_nil},
+           genstatem_opts
+         ) do
       {:ok, pid} when sync_connect_timeout == false ->
         {:ok, pid}
 
@@ -47,21 +52,24 @@ defmodule Xandra.Cluster.Pool do
     end
   end
 
-  @spec stop(GenServer.server(), term(), timeout()) :: :ok
+  @spec stop(:gen_statem.server_ref(), term(), timeout()) :: :ok
   def stop(pid, reason, timeout) do
-    GenServer.stop(pid, reason, timeout)
+    :gen_statem.stop(pid, reason, timeout)
   end
 
-  @spec checkout(GenServer.server()) :: {:ok, pid()} | {:error, :empty}
+  @spec checkout(:gen_statem.server_ref()) :: {:ok, pid()} | {:error, :empty}
   def checkout(pid) do
-    GenServer.call(pid, :checkout)
+    :gen_statem.call(pid, :checkout)
   end
 
-  ## State
+  ## Data
 
   defstruct [
     # Options for the underlying connection pools.
     :pool_options,
+
+    # Contact nodes.
+    :contact_nodes,
 
     # When auto-discovering nodes, you cannot get their port from C*.
     # Other drivers solve this by providing a static port that the driver
@@ -73,6 +81,10 @@ defmodule Xandra.Cluster.Pool do
 
     # The PID of the control connection.
     :control_connection,
+
+    # The interval to refresh the cluster's topology. Usually passed down
+    # to the control connection.
+    :refresh_topology_interval,
 
     # The load balancing policy info.
     :load_balancing_module,
@@ -92,19 +104,30 @@ defmodule Xandra.Cluster.Pool do
     # The name of the cluster (if present), only used for Telemetry events.
     :name,
 
-    # A map of peername to pool PID pairs.
-    pools: %{},
+    # A map of peername ({address, port}) to info about that peer.
+    # Each info map is:
+    # %{pool_pid: pid(), host: Host.t(), status: :up | :down | :connected}
+    peers: %{},
 
     # Modules to swap processes when testing.
     xandra_mod: nil,
     control_conn_mod: nil
   ]
 
+  ## Guards
+
+  defguardp is_peername(term)
+            when is_tuple(term) and tuple_size(term) == 2 and is_tuple(elem(term, 0)) and
+                   is_integer(elem(term, 1))
+
   ## Callbacks
 
   @impl true
+  def callback_mode, do: :handle_event_function
+
+  @impl true
   def init({cluster_opts, pool_opts, sync_connect_alias_or_nil}) do
-    {nodes, cluster_opts} = Keyword.pop!(cluster_opts, :nodes)
+    Process.flag(:trap_exit, true)
 
     registry_name =
       Module.concat([Xandra.ClusterRegistry, to_string(System.unique_integer([:positive]))])
@@ -126,206 +149,288 @@ defmodule Xandra.Cluster.Pool do
         {mod, opts} -> {mod, opts}
       end
 
-    state = %__MODULE__{
-      pool_options: pool_opts,
-      load_balancing_module: lb_mod,
-      load_balancing_state: lb_mod.init(lb_opts),
-      autodiscovered_nodes_port: Keyword.fetch!(cluster_opts, :autodiscovered_nodes_port),
-      xandra_mod: Keyword.fetch!(cluster_opts, :xandra_module),
-      control_conn_mod: Keyword.fetch!(cluster_opts, :control_connection_module),
-      target_pools: Keyword.fetch!(cluster_opts, :target_pools),
-      sync_connect_alias: sync_connect_alias_or_nil,
-      registry: registry_name,
-      name: Keyword.get(cluster_opts, :name),
-      pool_supervisor: pool_sup
-    }
-
-    {:ok, control_conn} =
-      state.control_conn_mod.start_link(
-        cluster: self(),
-        contact_points: nodes,
-        connection_options: state.pool_options,
-        autodiscovered_nodes_port: state.autodiscovered_nodes_port,
-        load_balancing: {lb_mod, lb_opts},
-        refresh_topology_interval: Keyword.fetch!(cluster_opts, :refresh_topology_interval),
+    data =
+      %__MODULE__{
+        pool_options: pool_opts,
+        contact_nodes: Keyword.fetch!(cluster_opts, :nodes),
+        load_balancing_module: lb_mod,
+        load_balancing_state: lb_mod.init(lb_opts),
+        autodiscovered_nodes_port: Keyword.fetch!(cluster_opts, :autodiscovered_nodes_port),
+        xandra_mod: Keyword.fetch!(cluster_opts, :xandra_module),
+        control_conn_mod: Keyword.fetch!(cluster_opts, :control_connection_module),
+        target_pools: Keyword.fetch!(cluster_opts, :target_pools),
+        sync_connect_alias: sync_connect_alias_or_nil,
         registry: registry_name,
-        cluster_name: Keyword.get(cluster_opts, :name)
-      )
+        name: Keyword.get(cluster_opts, :name),
+        pool_supervisor: pool_sup,
+        refresh_topology_interval: Keyword.fetch!(cluster_opts, :refresh_topology_interval)
+      }
 
-    state = %__MODULE__{state | control_connection: control_conn}
-
-    # This is only for testing purposes, and is not exposed.
-    if hosts = cluster_opts[:test_discovered_hosts] do
-      send(self(), {:discovered_hosts, hosts})
-      Enum.each(hosts, &send(self(), {:host_connected, &1}))
-    end
-
-    {:ok, state}
+    {:ok, @state, data, {:next_event, :internal, :start_control_connection}}
   end
 
   @impl true
-  def handle_call(:checkout, _from, %__MODULE__{} = state) do
-    {query_plan, state} =
-      get_and_update_in(state.load_balancing_state, fn lb_state ->
-        state.load_balancing_module.query_plan(lb_state)
+  def handle_event(type, event, state, data)
+
+  def handle_event(:internal, :start_control_connection, @state, data) do
+    case start_control_connection(data) do
+      {:ok, data} ->
+        {:keep_state, data}
+
+      :error ->
+        {:keep_state, data, {{:timeout, :reconnect_control_connection}, 1000, nil}}
+    end
+  end
+
+  def handle_event({:call, from}, :checkout, @state, %__MODULE__{} = data) do
+    {query_plan, data} =
+      get_and_update_in(data.load_balancing_state, fn lb_state ->
+        data.load_balancing_module.query_plan(lb_state)
       end)
 
     # Find the first host in the plan for which we have a pool.
     reply =
       query_plan
-      |> Stream.map(fn %Host{} = host -> Map.fetch(state.pools, Host.to_peername(host)) end)
-      |> Enum.find(_default = {:error, :empty}, &match?({:ok, _}, &1))
-
-    {:reply, reply, state}
-  end
-
-  @impl true
-  def handle_info(msg, state)
-
-  def handle_info({:host_up, %Host{} = host}, %__MODULE__{} = state) do
-    state = update_in(state.load_balancing_state, &state.load_balancing_module.host_up(&1, host))
-    state = maybe_start_pools(state)
-    {:noreply, state}
-  end
-
-  def handle_info({:host_connected, %Host{} = host}, %__MODULE__{} = state) do
-    state =
-      update_in(state.load_balancing_state, &state.load_balancing_module.host_connected(&1, host))
-
-    state = maybe_start_pools(state)
-
-    state =
-      if alias = state.sync_connect_alias do
-        send(alias, :connected)
-        %__MODULE__{state | sync_connect_alias: nil}
-      else
-        state
-      end
-
-    {:noreply, state}
-  end
-
-  def handle_info({:host_down, %Host{} = host}, %__MODULE__{} = state) do
-    state =
-      update_in(state.load_balancing_state, &state.load_balancing_module.host_down(&1, host))
-
-    state = stop_pool(state, host)
-    state = maybe_start_pools(state)
-    {:noreply, state}
-  end
-
-  def handle_info({:host_added, %Host{} = host}, %__MODULE__{} = state) do
-    state =
-      update_in(state.load_balancing_state, &state.load_balancing_module.host_added(&1, host))
-
-    state = maybe_start_pools(state)
-    {:noreply, state}
-  end
-
-  def handle_info({:host_removed, %Host{} = host}, %__MODULE__{} = state) do
-    state = stop_pool(state, host)
-
-    # Also delete the child from the supervisor altogether.
-    _ = Supervisor.delete_child(state.pool_supervisor, {host.address, host.port})
-
-    state =
-      update_in(state.load_balancing_state, &state.load_balancing_module.host_removed(&1, host))
-
-    state = update_in(state.pools, &Map.delete(&1, {host.address, host.port}))
-
-    state = maybe_start_pools(state)
-    {:noreply, state}
-  end
-
-  def handle_info({:discovered_hosts, hosts}, %__MODULE__{} = state) when is_list(hosts) do
-    execute_telemetry(state, [:discovered_peers], %{peers: hosts}, _extra_meta = %{})
-
-    state =
-      Enum.reduce(hosts, state, fn %Host{} = host, acc ->
-        update_in(acc.load_balancing_state, &state.load_balancing_module.host_added(&1, host))
+      |> Stream.map(fn %Host{} = host -> Map.fetch(data.peers, Host.to_peername(host)) end)
+      |> Enum.find_value(_default = {:error, :empty}, fn
+        {:ok, %{pool_pid: pid}} when is_pid(pid) -> {:ok, pid}
+        _other -> nil
       end)
 
-    state = maybe_start_pools(state)
-    {:noreply, state}
+    {:keep_state, data, {:reply, from, reply}}
+  end
+
+  def handle_event(:into, {:host_up, %Host{} = host}, @state, %__MODULE__{} = data) do
+    data = update_in(data.load_balancing_state, &data.load_balancing_module.host_up(&1, host))
+
+    data =
+      update_in(data.peers[Host.to_peername(host)], fn
+        %{status: :down} = peer -> %{peer | status: :up}
+        peer -> peer
+      end)
+
+    data = maybe_start_pools(data)
+    {:keep_state, data}
+  end
+
+  def handle_event(:info, {:host_down, %Host{} = host}, @state, %__MODULE__{} = data) do
+    data =
+      update_in(data.load_balancing_state, &data.load_balancing_module.host_down(&1, host))
+
+    data = put_in(data.peers[Host.to_peername(host)].status, :down)
+
+    data = stop_pool(data, host)
+    data = maybe_start_pools(data)
+    {:keep_state, data}
+  end
+
+  def handle_event(:info, {:discovered_hosts, new_peers}, @state, %__MODULE__{} = data)
+      when is_list(new_peers) do
+    execute_telemetry(data, [:discovered_peers], %{peers: new_peers}, _extra_meta = %{})
+
+    new_peers_map = Map.new(new_peers, &{Host.to_peername(&1), &1})
+    new_peers_set = MapSet.new(new_peers, &Host.to_peername/1)
+    old_peers_set = data.peers |> Map.keys() |> MapSet.new()
+
+    # Find the peers that are not in the set of known peers anymore, remove them
+    # from the LBP, emit a telemetry event, stop the pools.
+    data =
+      Enum.reduce(MapSet.difference(old_peers_set, new_peers_set), data, fn peername, data_acc ->
+        %{host: %Host{} = host} = Map.fetch!(data_acc.peers, peername)
+        handle_host_removed(data_acc, host)
+      end)
+
+    # For the new peers that we didn't know about, add them to the LBP, emit a
+    # telemetry event, and potentially start pools.
+    data =
+      Enum.reduce(MapSet.difference(new_peers_set, old_peers_set), data, fn peername, data_acc ->
+        handle_host_added(data_acc, Map.fetch!(new_peers_map, peername))
+      end)
+
+    {:keep_state, data}
+  end
+
+  # Sent by the connection itself.
+  def handle_event(
+        :info,
+        {:xandra, :connected, peername, _pid},
+        @state,
+        %__MODULE__{} = data
+      )
+      when is_peername(peername) do
+    data = put_in(data.peers[peername].status, :connected)
+
+    host = data.peers[peername].host
+
+    data =
+      update_in(data.load_balancing_state, &data.load_balancing_module.host_connected(&1, host))
+
+    data =
+      if alias = data.sync_connect_alias do
+        send(alias, :connected)
+        %__MODULE__{data | sync_connect_alias: nil}
+      else
+        data
+      end
+
+    {:keep_state, data}
+  end
+
+  # Sent by the connection itself.
+  def handle_event(
+        :info,
+        {:xandra, :failed_to_connect, peername, _pid},
+        @state,
+        %__MODULE__{} = data
+      ) do
+    data = put_in(data.peers[peername].status, :down)
+    data = stop_pool(data, data.peers[peername].host)
+    {:keep_state, data}
+  end
+
+  def handle_event({:timeout, :reconnect_control_connection}, nil, @state, %__MODULE__{} = data) do
+    {:keep_state, data, {:next_event, :internal, :start_control_connection}}
   end
 
   ## Helpers
 
+  defp handle_host_added(%__MODULE__{} = data, %Host{} = host) do
+    data =
+      update_in(data.load_balancing_state, &data.load_balancing_module.host_added(&1, host))
+
+    data =
+      update_in(
+        data.peers,
+        &Map.put(&1, Host.to_peername(host), %{host: host, status: :up, pool_pid: nil})
+      )
+
+    execute_telemetry(data, [:change_event], %{}, %{event_type: :host_added, host: host})
+
+    maybe_start_pools(data)
+  end
+
+  defp handle_host_removed(%__MODULE__{} = data, %Host{} = host) do
+    data =
+      update_in(data.load_balancing_state, &data.load_balancing_module.host_removed(&1, host))
+
+    data = stop_pool(data, host)
+    _ = Supervisor.delete_child(data.pool_supervisor, Host.to_peername(host))
+
+    data = update_in(data.peers, &Map.delete(&1, Host.to_peername(host)))
+
+    execute_telemetry(data, [:change_event], %{}, %{event_type: :host_removed, host: host})
+
+    data
+  end
+
   # This function is idempotent: you can call it as many times as you want with the same
   # peer, and it'll only start it once.
-  defp start_pool(state, %Host{} = host) do
+  defp start_pool(%__MODULE__{} = data, %Host{} = host) do
     conn_options =
-      Keyword.merge(state.pool_options,
+      Keyword.merge(data.pool_options,
         nodes: [Host.format_address(host)],
-        registry: state.registry,
-        connection_listeners: [state.control_connection]
+        registry: data.registry,
+        cluster_pid: self()
       )
 
     peername = Host.to_peername(host)
 
     pool_spec =
-      Supervisor.child_spec({state.xandra_mod, conn_options},
-        id: peername,
-        restart: :transient
-      )
+      Supervisor.child_spec({data.xandra_mod, conn_options}, id: peername, restart: :transient)
 
-    case Supervisor.start_child(state.pool_supervisor, pool_spec) do
+    case Supervisor.start_child(data.pool_supervisor, pool_spec) do
       {:ok, pool} ->
-        execute_telemetry(state, [:pool, :started], %{}, %{host: host})
-        send(state.control_connection, {:started_pool, host})
-        put_in(state.pools[peername], pool)
+        execute_telemetry(data, [:pool, :started], %{}, %{host: host})
+        put_in(data.peers[peername].pool_pid, pool)
 
       {:error, :already_present} ->
-        case Supervisor.restart_child(state.pool_supervisor, _id = peername) do
+        case Supervisor.restart_child(data.pool_supervisor, _id = peername) do
           {:ok, pool} ->
-            execute_telemetry(state, [:pool, :restarted], %{}, %{host: host})
-            send(state.control_connection, {:started_pool, host})
-            put_in(state.pools[peername], pool)
+            execute_telemetry(data, [:pool, :restarted], %{}, %{host: host})
+            put_in(data.peers[peername].pool_pid, pool)
 
           {:error, reason} when reason in [:running, :restarting] ->
-            state
+            data
 
           {:error, other} ->
             raise "unexpected error when restarting pool for #{Host.format_address(host)}: #{inspect(other)}"
         end
 
       {:error, {:already_started, _pool}} ->
-        state
+        data
     end
   end
 
-  defp stop_pool(state, %Host{} = host) do
-    _ = Supervisor.terminate_child(state.pool_supervisor, {host.address, host.port})
-    update_in(state.pools, &Map.delete(&1, {host.address, host.port}))
+  defp stop_pool(data, %Host{} = host) do
+    peername = Host.to_peername(host)
+
+    execute_telemetry(data, [:pool, :stopped], %{}, %{host: host})
+    _ = Supervisor.terminate_child(data.pool_supervisor, peername)
+
+    put_in(data.peers[peername].pool_pid, nil)
   end
 
-  defp maybe_start_pools(%__MODULE__{target_pools: target, pools: pools} = state)
-       when map_size(pools) == target do
-    state
+  defp maybe_start_pools(%__MODULE__{target_pools: target} = data) do
+    if Enum.count(data.peers, fn {_peername, %{pool_pid: pid}} -> is_pid(pid) end) == target do
+      data
+    else
+      {hosts_plan, data} =
+        get_and_update_in(data.load_balancing_state, &data.load_balancing_module.hosts_plan/1)
+
+      Enum.reduce_while(hosts_plan, data, fn %Host{} = host, state ->
+        case Map.fetch!(data.peers, Host.to_peername(host)) do
+          %{pool_pid: pid} when is_pid(pid) ->
+            {:cont, state}
+
+          %{pool_pid: nil} ->
+            data = start_pool(data, host)
+
+            if Enum.count(data.peers, fn {_peername, %{pool_pid: pid}} -> is_pid(pid) end) ==
+                 target do
+              {:halt, data}
+            else
+              {:cont, data}
+            end
+        end
+      end)
+    end
   end
 
-  defp maybe_start_pools(%__MODULE__{target_pools: target, pools: pools} = state)
-       when map_size(pools) < target do
-    {hosts_plan, state} =
-      get_and_update_in(state.load_balancing_state, fn lb_state ->
-        state.load_balancing_module.hosts_plan(lb_state)
+  defp start_control_connection(%__MODULE__{} = data) do
+    {lbp_hosts, data} =
+      get_and_update_in(data.load_balancing_state, fn lb_state ->
+        data.load_balancing_module.query_plan(lb_state)
       end)
 
-    Enum.reduce_while(hosts_plan, state, fn %Host{} = host, state ->
-      case Map.fetch(pools, Host.to_peername(host)) do
-        {:ok, _pool} ->
-          {:cont, state}
+    contact_nodes_as_hosts =
+      Enum.map(data.contact_nodes, fn {address, port} ->
+        %Host{address: address, port: port}
+      end)
 
-        :error ->
-          state = start_pool(state, host)
+    case start_control_connection(data, lbp_hosts ++ contact_nodes_as_hosts) do
+      {:ok, control_conn} -> {:ok, %__MODULE__{data | control_connection: control_conn}}
+      :error -> :error
+    end
+  end
 
-          if map_size(state.pools) == target do
-            {:halt, state}
-          else
-            {:cont, state}
-          end
-      end
-    end)
+  defp start_control_connection(_data, []) do
+    :error
+  end
+
+  defp start_control_connection(data, [%Host{} = host | hosts]) do
+    control_conn_opts = [
+      cluster_pid: self(),
+      cluster_name: data.name,
+      contact_node: {host.address, host.port},
+      connection_options: data.pool_options,
+      autodiscovered_nodes_port: data.autodiscovered_nodes_port,
+      refresh_topology_interval: data.refresh_topology_interval
+    ]
+
+    case data.control_conn_mod.start_link(control_conn_opts) do
+      {:ok, control_conn} -> {:ok, control_conn}
+      {:error, _reason} -> start_control_connection(data, hosts)
+    end
   end
 
   defp execute_telemetry(%__MODULE__{} = state, event_postfix, measurements, extra_meta) do
